refactor(indexing): move embedding/search hooks from lib/settings to lib/indexing
These hooks fetch indexing configuration (embedding models, reranking,
provider credentials), not application settings — they belong alongside the
interfaces they consume in lib/indexing/.

- Create lib/indexing/hooks.ts with secondaryRefreshInterval,
  useCurrentSearchSettings, useSecondarySearchSettings,
  useCurrentEmbeddingModel, useLLMContextualCosts,
  useConfiguredEmbeddingProviders
- Trim lib/settings/hooks.ts to its true scope: SettingsContext,
  useSettingsContext, useVectorDbEnabled
- Update IndexSettingsPage and useSearchSettings.test.ts to import from
  @/lib/indexing/hooks
