+++ web/src/hooks/useSearchSettings.test.ts
@@ -1,4 +1,4 @@
-import { secondaryRefreshInterval } from "@/lib/settings/hooks";
+import { secondaryRefreshInterval } from "@/lib/indexing/hooks";
 import type { EmbeddingModelResponse } from "@/lib/indexing/interfaces";
 
 const embeddingModelResponse: EmbeddingModelResponse = {
