+++ web/src/lib/settings/hooks.ts
@@ -1,33 +1,15 @@
 "use client";
 
 /**
- * Settings-related React hooks.
+ * Application settings hooks.
  *
- * Two categories:
- *   1. Context readers — `useSettingsContext`, `useVectorDbEnabled`. These
- *      read from the `SettingsContext` that `SettingsProvider` populates.
- *      The vast majority of components use these.
- *   2. Search / embedding fetchers — `useCurrentSearchSettings`,
- *      `useSecondarySearchSettings`, `useCurrentEmbeddingModel`,
- *      `useLLMContextualCosts`, `useConfiguredEmbeddingProviders`. These are
- *      used primarily by the admin Index Settings page.
- *
- * The SWR fetchers that back `SettingsProvider` (`useSettings`,
- * `useEnterpriseSettings`, `useCustomAnalyticsScript`) are private to
- * `providers/SettingsProvider.tsx` and are not exported here.
+ * Provides access to the `CombinedSettings` context populated by
+ * `SettingsProvider`. For embedding model / indexing configuration hooks,
+ * see `lib/indexing/hooks.ts`.
  */
 
 import { createContext, useContext } from "react";
-import useSWR from "swr";
-import { errorHandlingFetcher } from "@/lib/fetcher";
-import { SWR_KEYS } from "@/lib/swr-keys";
 import { CombinedSettings } from "@/interfaces/settings";
-import {
-  ConfiguredEmbeddingProvider,
-  EmbeddingModelResponse,
-  LLMContextualCost,
-  SavedSearchSettings,
-} from "@/lib/indexing/interfaces";
 
 // ─── Settings Context ───────────────────────────────────────────────────────
 
@@ -73,89 +55,4 @@ export function useSettingsContext() {
 export function useVectorDbEnabled(): boolean {
   const settings = useSettingsContext();
   return settings.settings.vector_db_enabled !== false;
-}
-
-// ─── Search / Embedding Hooks ───────────────────────────────────────────────
-
-/**
- * Determines the SWR `refreshInterval` for the secondary (in-progress)
- * embedding model poll:
- * - 5 s while a migration is in flight (data present)
- * - 60 s otherwise (no migration running — catch migrations started elsewhere)
- */
-export function secondaryRefreshInterval(
-  latestData: EmbeddingModelResponse | null | undefined
-): number {
-  return latestData ? 5000 : 60000;
-}
-
-/**
- * Fetch the active search / embedding settings.
- * Polls only when `pollIntervalMs` is provided.
- */
-export function useCurrentSearchSettings({
-  pollIntervalMs = 0,
-}: { pollIntervalMs?: number } = {}) {
-  return useSWR<SavedSearchSettings | null>(
-    SWR_KEYS.currentSearchSettings,
-    errorHandlingFetcher,
-    { refreshInterval: pollIntervalMs }
-  );
-}
-
-/**
- * Fetch the secondary (in-progress) embedding model.
- * Returns `null` when no re-index is running.
- * Self-throttles via `secondaryRefreshInterval`.
- */
-export function useSecondarySearchSettings() {
-  return useSWR<EmbeddingModelResponse | null>(
-    SWR_KEYS.secondarySearchSettings,
-    errorHandlingFetcher,
-    { refreshInterval: secondaryRefreshInterval }
-  );
-}
-
-/**
- * Fetch the active embedding model from the current search settings key.
- * Polls only when `pollIntervalMs` is provided.
- *
- * The returned shape does NOT carry a `description` — descriptions are
- * frontend-only. Look them up via `getCurrentModelCopy` if needed.
- */
-export function useCurrentEmbeddingModel({
-  pollIntervalMs = 0,
-}: { pollIntervalMs?: number } = {}) {
-  return useSWR<EmbeddingModelResponse | null>(
-    SWR_KEYS.currentSearchSettings,
-    errorHandlingFetcher,
-    { refreshInterval: pollIntervalMs }
-  );
-}
-
-/**
- * Fetch LLM models available for contextual RAG, including per-model token
- * cost.
- */
-export function useLLMContextualCosts() {
-  return useSWR<LLMContextualCost[]>(
-    SWR_KEYS.llmContextualCost,
-    errorHandlingFetcher
-  );
-}
-
-/**
- * Fetch cloud embedding providers that have credentials configured in the
- * backend.
- *
- * Returns a plain array rather than a `Map` — SWR's internal hash comparison
- * doesn't reliably detect changes between two `Map` instances, which caused
- * stale views after `mutate`. Build a lookup `Map` client-side via `useMemo`
- * if needed.
- */
-export function useConfiguredEmbeddingProviders() {
-  return useSWR<ConfiguredEmbeddingProvider[]>(
-    SWR_KEYS.embeddingProviders,
-    errorHandlingFetcher
-  );
 }
